HBAI-227 Move embedding code to own route
